+++ CHANGELOG.md
@@ -6,6 +6,9 @@ most recent changes however.
 
 ## April
 
+- Rye's `add` command now accepts a `--pre` parameter to include pre-release
+  candidates.
+
 - Rye's `pin` command now updates the pyproject.toml requires-python.
 
 - Rye's `install` command now accepts a `--include-dep` parameter to include